Initial commit of the plugin
In this commit, we have initially added abandoned cart data ( product name & quantity ) to the agile CRM. When This plugin is activated it will create a new tab in the abandoned menu page. In the New created Tab we have given 3 settings named as 1. Export abandoned cart data to Agile CRM 2.Automatically add abandoned cart data to Agile CRM 3. Automatically add abandoned cart data to Agile CRM after a set time.

If setting named as Export abandoned cart data to Agile CRM is enable then only it will display 3 methods of the exporting the data to Agile CRM. 1. Single cart ( hover effect ) 2. Bulk action 3. Export all data.

If setting named as 'Automatically add abandoned cart data to Agile CRM' is enabled then automatically Cron job will add the abandoned carts data to the agile CRM. If this setting is not enabled then automatically abandoned cart will not be added to the agile CRM. Admin can set the on job time using setting 'Automatically add abandoned cart data to Agile CRM after set time.'

When any data is sent to the agile CRM it will create the 'Customer' and the abandoned cart products and its quantity will be added as a NOTES. Please note: Agile CRM keep the track of the customer by their EMAIL address. So each user has the single record for the Agile CRM, but the abandoned carts data will be added in the NOTES section for the same users.