Created test connection
We have created one setting under which customer can check the test connection of the given setting

There is a condition in the test connection. If the Test connection is successfully established, then while we click on the save changes button, it should not again check for the test connection. The directly setting must be saved.

If Customer has not tested the test connection than when customer click on the Save changes button our plugin will automatically check the connection. If the connection is not established successfully then Setting will not be SAVED.

If connection established successfully then only settings will be saved.

NOTE: You need to update the Abandoned cart pro plugin also. Please take the updated plugin only.