Fix #24
This fix ensure that when we export the data from the BULK edit drop down it will allow us to add the data. If the data is already exported then it will show the proper data. It will not show the message of the visitors cart until you have selected only visitors carts